Change the NS_PER_SEC (and friends) from enum to static const
New version of clang (19) has introduced a stricter checks when mixing
integer (and float types) with enums.  In this case, we used enum {}
as C17 doesn't have constexpr yet.  Change the time conversion constants
to be static const unsigned int instead of enum values.

(cherry picked from commit b03e90e0d4d463a0b729d0333f99a3535a9f6a5c)

+++ lib/isc/stdtime.c
@@ -40,7 +40,8 @@ isc_stdtime_now(void) {
 	if (clock_gettime(CLOCKSOURCE, &ts) == -1) {
 		FATAL_SYSERROR(errno, "clock_gettime()");
 	}
-	INSIST(ts.tv_sec > 0 && ts.tv_nsec >= 0 && ts.tv_nsec < NS_PER_SEC);
+	INSIST(ts.tv_sec > 0 && ts.tv_nsec >= 0 &&
+	       ts.tv_nsec < (long)NS_PER_SEC);
 
 	return ((isc_stdtime_t)ts.tv_sec);
 }
